Feat/management distribution meeting analysis (#14)
* feat: management distribution

* fix: prompt selector for roles

* first commit for meeting analysis

* working distribution, but no meeting analysis yet

* prototype files

* prompts

* dff_meeting_analysis_skill instead of prompted; llm-based everything

* working version of meeting analysis skill

* dff_meeting_analysis_skill with 4 nodes

* doc-processor annotator

* added saving previous meeting analysis results; links to them are  written to bot attrs

* update roles

* fix for meeting analysis skill, now working

* document only for now, then will be deleted

* prompt for unabridged response selection

* refactor doc_processor, remove unnecessary funcs

* better prompts

* better skill description in components

* add llm-based-skill-selector to dist

* enable finding previously generated meeting analyses; better fallback

* 512 max_tokens for chatgpt in some cases

* enhance response selector prompt

* add dff_meeting_analysis_skill_formatter

* some fixes to cards and configs

* update readmes

* correct ports for doc processor; remove extra prompt

* codestyle

* codestyle

* fixes for Dilya

* enhanced checks

* typo

* codestyle + small fix for checks

* file moved to google drive

* remove extra print

* checking each file if processed; concatenating multiple files; two containers for doc-processor

* typo fix

* unique ids for files in data/, ids to paths in config

* delete transcript files

* codestyle

* fix: UIDs for files in data now working

* fixes in working with files

* codestyle

* fix error in getting related_files

* Revert "fix error in getting related_files"

This reverts commit 705e23897e9317e1ba24702b14e7c097da093dcd.

* working fix for bot_attrs_files

* remove document file

* numerous fixes for review

* codestyle

* bring some things to common

* even better funcs in common

* codestyle

* saving all processed docs in atts; saving candidate texts in adds of utt; link or path possible for processing from atts

* fixes for accidentally broken stuff

* some more fixes

* candidate texts to hyp attributes

* codestyle

* FILE_SERVER_TIMEOUT as arg

* GENERATIVE_SERVICE_URL as arg

* fix: formatters in pipeline_conf

* component card for vectorize_documents

* openai-chatgpt-long.json for document-qa-llm-skill

* openai-chatgpt-long.json for meeting-analysis-skill

* fix: timeouts and component card paths

* add regex for http check

* doc processor names in service_config files

* update getting envvars

* codestyle

* fix: remove envvars from everywhere

* fix: remove envvars from everywhere

* fixes: details in cards and pipeline

* fixes: details in cards and pipeline

* feat: special message if failed to process file from atts

* get token limit from service endpoint

* fix: better upload_document, try except inside func & enable both text and file upload in one func

* docstrings; also fix: detecting extension for links

* codestyle

* again codestyle

* update READMEs with dialog state info

* fix: add diff endpoints to doc-retriever readme

* fix: solve inconsistencies in cards and readmes

* fix: incorrect formatters in cards

* update ports to non-allocated ones

* fixes: everything acc to comments

* codestyle

* generalize file service url in another comment

* codestyle

* refactor attributes structure

* update readmes to include info about new attributes format

* fix: clean config; comment about format

* add comments; {FILE_SERVER_URL} instead of actual url

* comments and readmes

* implement storing doc for N_TURNS_FOR_DISCUSSION turns

* codestyle

* improve N_TURNS_FOR_DISCUSSION, implement only for doc-processor-from-atts

* better logging in doc-retriever

* codestyle

* more comments

* codestyle

* delete extra logs

* some more comments

* count n_steps_discussed in any case; put that to readme

* fix: n_steps_discussed in correct place

* fix: if file was processed earlier, take processed text from processed_documents

* if we get doc from somewhere, consider it good as new -> reset n_steps_discussed to 0

* codestyle

* update comments; fix logic of n_steps_discussed

* better comments

* fix: small fixes

* N_TURNS_FOR_DISCUSSION: -> N_TURNS_TO_KEEP_DOC

* N_TURNS_TO_KEEP_DOC in distribution files

* N_TURNS_TO_KEEP_DOC: 10 ->; also updates in readmes and comments

* codestyle

* comment about N_TURNS_TO_KEEP_DOC

* comment about N_TURNS_TO_KEEP_DOC

* fix: remove sentseg from management dist

* better descriptions for skills

* fix hyp format for dff_meeting_analysis_skill

* fixes: remove logs, improve skill description

* ensure unique ids everywhere; add dialog_id to file_id

* update skill selector: turn off doc-based skills when we don't have doc

* codestyle

* codestyle again

* remove one extra log

* now we can also process files from file server

* codestyle

* fix: is_container_running to response.py

* fix to prompts; also longer context for many services

* always turn on document-based wa skill

* codestyle

* add file exists check

* start adding question_answering default node

* node for question answering in meeting analysis skill; small change in llm-based-skill-selector

* codestyle

* condition file

* Dilya's fixes for skill-selector

* codestyle

* slightly improve prompt for response selector

* fix: chunks only split by newlines

* fix: no extra info in prompts; better response selector

* small fixes

* codestyle

* added list title

* codestyle

* codestyle

* moved is_container_running up

* fix: tags: selector

* add check if skill to add is in pipeline

* shorter prompt for response selector

* copy older dist with tf-idf qa as management_assistant_extended

* remove tf-idf qa skill from management assistant

* update description for meeting analysis skill

* remover doc-retriever from main distribution

* better guidance for qa

* feat: turn on dff_meeting_analysis_skill when it was used with the same doc before

* codestyle

* codestyle

* fix: only perform doc-related checks in skill selector if we actually have a doc in use

* fix: include situation when we don't have prev_skills or prev_docs in skill selector

* fix: only check last_n_active_skills_with_same_doc if we have N_TURNS_TO_KEEP_DOC

* missing service config files

* fixes acc to Dilya's review

* codestyle

* fixes acc to Ksusha's review

* CODESTYLE

* formatter rename and retrun older cropped dialog

* extended dff_prompted_skill_formatter, now no need for separate meeting_analysis_formatter

* replace cropped_dialog with cropped_dialog_skills_and_docs in all llm-based-skill-selectors

* fix: remove quotes

* fix in skill selector for Dilya; correct formatter in doc_processor

* fix: required services

* some more fixes in llm based skill selector

* fix typo

---------

+++ skills/dff_document_qa_llm_skill/scenario/response.py
@@ -5,9 +5,10 @@ import re
 import requests
 import sentry_sdk
 import time
-from typing import Any, List
+import copy
 
-from common.build_dataset import build_dataset
+from typing import Any
+from common.build_dataset import build_dataset, get_text_for_candidates
 import common.dff.integration.context as int_ctx
 import common.dff.integration.response as int_rsp
 from common.constants import CAN_NOT_CONTINUE
@@ -62,40 +63,35 @@ with open(DOCUMENT_PROMPT_FILE, "r") as f:
 
 
 FIX_PUNCTUATION = re.compile(r"\s(?=[\.,:;])")
-FIND_ID = re.compile(r"file=([0-9a-zA-Z]*).txt")
+FIND_ID = re.compile(r"file=([0-9a-zA-Z_]*).txt")
 DEFAULT_CONFIDENCE = 0.9
 SUPER_CONFIDENCE = 1.0
 LOW_CONFIDENCE = 0.7
 
 
-def get_text_for_candidates(dataset_path: str, raw_candidates: List[str]) -> str:
-    num_candidates = []
-    nums = 0
-    for f_name in raw_candidates:
-        nums += 1
-        with open(dataset_path + f_name) as f:
-            num_candidates.append(f"{nums}. {f.read()}")
-    return " ".join(num_candidates)
-
-
 def compose_data_for_model(ctx: Context, actor: Actor) -> str:
     if not os.path.exists("/data/documents"):
         os.mkdir("/data/documents")
+    filepaths_on_server, filepaths_in_container = [], []
+    final_candidates = ""
     dialog = int_ctx.get_dialog(ctx, actor)
     context = dialog.get("utterances", [])[-N_UTTERANCES_CONTEXT:]
     utterance_texts = [uttr.get("text", "") for uttr in context]
+    utterances_with_doc_text = copy.deepcopy(utterance_texts)
     if utterance_texts:
         raw_candidates = context[-1].get("annotations", {}).get("doc_retriever", {}).get("candidate_files", [])
-        filepaths_on_server = (
-            context[-1].get("user", {}).get("attributes", {}).get("documents_qa_model", {}).get("document_links", [])
-        )
-        filepaths_in_container = []
+        processed_docs = context[-1].get("user", {}).get("attributes", {}).get("documents_in_use", {})
+        if processed_docs:
+            filepaths_on_server = [
+                processed_docs[key].get("full_processed_text_link", "") for key in processed_docs.keys()
+            ]
         for filepath in filepaths_on_server:
             file_id = re.search(FIND_ID, filepath).group(1)
             filepath_container = f"/data/documents/{file_id}.txt"
-            orig_file = requests.get(filepath, timeout=FILE_SERVER_TIMEOUT)
-            with open(filepath_container, "wb") as f:
-                f.write(orig_file.content)
+            if not os.path.exists(filepath_container):  # if we don't have doc in container, download it from server
+                orig_file = requests.get(filepath, timeout=FILE_SERVER_TIMEOUT)
+                with open(filepath_container, "wb") as f:
+                    f.write(orig_file.content)
             filepaths_in_container.append(filepath_container)
         dataset_path = "/data/temporary_dataset/"
         if not os.path.exists(dataset_path):
@@ -107,9 +103,9 @@ filepaths_in_container: {filepaths_in_container}, dataset_path: {dataset_path}""
         build_dataset(dataset_path, filepaths_in_container)
         logger.info("Dataset built successfully")
         final_candidates = get_text_for_candidates(dataset_path, raw_candidates)
-        request = utterance_texts[-1]
-        utterance_texts[-1] = f"""Text: ### {final_candidates} ###\nUSER: {request}\n{DOCUMENT_PROMPT_TEXT}"""
-    return utterance_texts
+        request = utterances_with_doc_text[-1]
+        utterances_with_doc_text[-1] = f"""Text: ### {final_candidates} ###\n{DOCUMENT_PROMPT_TEXT}\nUser: {request}"""
+    return utterances_with_doc_text, utterance_texts, final_candidates
 
 
 def generative_response(ctx: Context, actor: Actor, *args, **kwargs) -> Any:
@@ -130,7 +126,8 @@ def generative_response(ctx: Context, actor: Actor, *args, **kwargs) -> Any:
             curr_bot_attrs += [bot_attr]
             curr_attrs += [attr]
 
-    dialog_context = compose_data_for_model(ctx, actor)
+    dialog_context_to_send, dialog_context_for_logging, candidate_texts = compose_data_for_model(ctx, actor)
+    hyp_attrs = {"candidate_chunks_text": candidate_texts}  # saving raw texts of candidate answers to hyp attributes
     human_uttr_attributes = int_ctx.get_last_human_utterance(ctx, actor).get("attributes", {})
     envvars_to_send = ENVVARS_TO_SEND if len(ENVVARS_TO_SEND) else human_uttr_attributes.get("envvars_to_send", [])
     sending_variables = compose_sending_variables(
@@ -138,12 +135,12 @@ def generative_response(ctx: Context, actor: Actor, *args, **kwargs) -> Any:
         envvars_to_send,
         human_uttr_attributes,
     )
+    logger.info(f"dialog_context: {dialog_context_for_logging}")
 
-    logger.info(f"dialog_context: {dialog_context}")
-    if len(dialog_context) > 0:
+    if len(dialog_context_to_send) > 0:
         try:
             hypotheses = send_request_to_prompted_generative_service(
-                dialog_context,
+                dialog_context_to_send,
                 DEFAULT_SYSTEM_PROMPT,
                 GENERATIVE_SERVICE_URL,
                 GENERATIVE_SERVICE_CONFIG,
@@ -164,6 +161,7 @@ def generative_response(ctx: Context, actor: Actor, *args, **kwargs) -> Any:
             hyp += "."
             confidence = LOW_CONFIDENCE
         _curr_attrs = {"can_continue": CAN_NOT_CONTINUE}
+        _curr_attrs.update(hyp_attrs)
         gathering_responses(hyp, confidence, {}, {}, _curr_attrs)
 
     if len(curr_responses) == 0:
